+++ source/Internal/Framework/Module/Setup/Service/ModuleConfigurationHandlingService.php
@@ -10,57 +10,28 @@ declare(strict_types=1);
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Service;
 
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Handler\ModuleConfigurationHandlerInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Validator\ModuleConfigurationValidatorInterface;
 
 class ModuleConfigurationHandlingService implements ModuleConfigurationHandlingServiceInterface
 {
     /**
-     * @var ModuleConfigurationHandlerInterface[]
-     */
-    private $handlers = [];
-
-    /**
      * @var ModuleConfigurationValidatorInterface[]
      */
-    private $moduleConfigurationValidators = [];
+    private array $moduleConfigurationValidators = [];
 
     /**
      * @param ModuleConfiguration $moduleConfiguration
      * @param int                 $shopId
      */
-    public function handleOnActivation(ModuleConfiguration $moduleConfiguration, int $shopId)
+    public function handleOnActivation(ModuleConfiguration $moduleConfiguration, int $shopId): void
     {
         $this->validateModuleConfiguration($moduleConfiguration, $shopId);
-
-        foreach ($this->handlers as $handler) {
-            $handler->handleOnModuleActivation($moduleConfiguration, $shopId);
-        }
-    }
-
-    /**
-     * @param ModuleConfiguration $moduleConfiguration
-     * @param int                 $shopId
-     */
-    public function handleOnDeactivation(ModuleConfiguration $moduleConfiguration, int $shopId)
-    {
-        foreach ($this->handlers as $handler) {
-            $handler->handleOnModuleDeactivation($moduleConfiguration, $shopId);
-        }
-    }
-
-    /**
-     * @param ModuleConfigurationHandlerInterface $moduleSettingHandler
-     */
-    public function addHandler(ModuleConfigurationHandlerInterface $moduleSettingHandler)
-    {
-        $this->handlers[] = $moduleSettingHandler;
     }
 
     /**
      * @param ModuleConfigurationValidatorInterface $configuration
      */
-    public function addValidator(ModuleConfigurationValidatorInterface $configuration)
+    public function addValidator(ModuleConfigurationValidatorInterface $configuration): void
     {
         $this->moduleConfigurationValidators[] = $configuration;
     }
@@ -69,7 +40,7 @@ class ModuleConfigurationHandlingService implements ModuleConfigurationHandlingS
      * @param ModuleConfiguration $moduleConfiguration
      * @param int                 $shopId
      */
-    private function validateModuleConfiguration(ModuleConfiguration $moduleConfiguration, int $shopId)
+    private function validateModuleConfiguration(ModuleConfiguration $moduleConfiguration, int $shopId): void
     {
         foreach ($this->moduleConfigurationValidators as $moduleConfigurationValidator) {
             $moduleConfigurationValidator->validate($moduleConfiguration, $shopId);